Parse Firebase key from the .env file
Instead of using the Firebase key JSON file directly in the project
directory, this change reads from the .env file, and it is parsed from
string to an object. This way it will be easier to use external tools
in the production environment to bring the JSON key to the code.