Move all trade actions to viewtrade.js
The trade functionality has been consolidated into a single command to
simplify the trading process. The commands for accepting, canceling, and
declining trade requests have been removed, and their logic has been
integrated into the 'viewtrade.js' file, with buttons.

Also, the logic for the button handling and displaying were correctly
implemented. It will just show the appropiate buttons to the issuer and
recipient, if the request is still pending.