Fix cron schedule for market update and crystals
The cron schedule for the market update and crystal distribution has
been corrected. The market update resets every Monday instead of Sunday,
and the crystal distribution occurs at the beginning of the month.

The information commands involved in this change are `faq.js` and
`vip.js`. Also the `README.md` was updated to reflect these changes.