Show the price of each card on the market
The price of each card is shown on the market image. The price is also
displayed more transparent if the card was already purchased or if the
user does not have enough crystals to buy it.

Some components were resized (position only) so that prices could be
fitted in the image without looking overloaded.